+++ pywren_ibm_cloud/version.py
@@ -1,5 +1,5 @@
 #
-# (C) Copyright IBM Corp. 2018
+# (C) Copyright IBM Corp. 2020
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
